Fix progressPercentage format and CategoryResponse custom field for E2E tests

+++ src/main/java/com/finance/manager/dto/response/GoalResponse.java
@@ -17,6 +17,6 @@ public class GoalResponse {
     private LocalDate targetDate;
     private LocalDate startDate;
     private BigDecimal currentProgress;
-    private double progressPercentage;
+    private BigDecimal progressPercentage;
     private BigDecimal remainingAmount;
 }
